+++ src/app/batteries/batteries.component.ts
@@ -5,7 +5,7 @@ import { MatTableModule } from '@angular/material/table';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
-import { DataRecordStore } from '../store/data-record-store';
+import { DataRecordStore } from '../store/csv-data-loading/data-record-store';
 import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -43,7 +43,6 @@ export class BatteriesComponent {
 
   addBattery() {
     if (this.batteryForm.valid) {
-
       const newBattery = this.batteryForm.value as Battery;
       // this.store.addBattery(newBattery);
       console.log('New Battery:', newBattery);
