+++ src/app/tariff-rates/tariff-rates.component.html
@@ -1,23 +1,3 @@
-<!-- @for(tariffPrice of tariffPrices; track tariffPrice.name){
-
-<mat-card>
-  <mat-card-header>
-    <mat-card-title>{{ tariffRate.name }}</mat-card-title>
-  </mat-card-header>
-  <mat-card-content>
-    @for(line of tariffRate.lines; track line.tariffName){
-
-    <p>
-      {{ line.tariffName }} cena: {{ line.values.total | number : "1.2-2" }} zł
-      (prąd {{ line.values.total | number : "1.2-2" }} zł dystrybucja
-      {{ line.values.distribution | number : "1.2-2" }} zł)
-    </p>
-
-    }
-  </mat-card-content>
-</mat-card>
-} -->
-
 @for(tariffPrice of tariffPrices; track tariffPrice.tariff){
 <mat-card>
   <mat-card-header>
@@ -25,9 +5,7 @@
   </mat-card-header>
   <mat-card-content>
     @for(zone of tariffPrice.zones; track zone.name){
-    <p>
-      {{ zone | json }}
-    </p>
+    <p>{{ zone.name }}: {{ zone.price.total | number : "1.2-2" }} zł</p>
     }
   </mat-card-content>
 </mat-card>
